requirements.txt
google-genai
structlog